// src/components/interfaces.tsx
export interface People {
  birth_year: string;
  created: string;
  eye_color: string;
  films: string[];
  gender: string;
  hair_color: string;
  height: string;
  homeworld: string;
  mass: string;
  name: string;
  skin_color: string;
  species: string[];
  starships: string[];
  url: string;
  edited: string;
  vehicles: string[];
}